Checkpoint Update
Added checkpoints. They are tracked as a new variable in Session.java. A
command from the server sets the player's checkpoint. Appropriate parser
has been added. Adding a checkpoint replaces the old one from the
database, appropriate database variables and methods have been added.

The plugin has the ability to turn off joining and only allow the
viewing of records. This is disabled by the boolean variable "dungeons"
in the config.

Added functionality for viewing the history of records in dungeons.